16/03/09

+++ WebRoot/jsp/success.jsp
@@ -122,38 +122,34 @@
 		id="frmfindId">
 		<table align="center">
 			<tr>
-				<td><input name="tbOrder" id="txtOrder" value="${order }" />
+				<td><input type="hidden" name="tbOrder" id="txtOrder"
+					value="${order }" /> <input type="hidden" name="page" value="1" />
 				</td>
 				<td align="right">收信人</td>
 				<td align="left"><input name="txtRecipient" type="text"
-					style="width: 180px;" value="${recipient }" />
-				</td>
+					style="width: 180px;" value="${recipient }" /></td>
 				<td align="right" style="h1">发送日期</td>
 				<td align="left"><input name="txtDateStart" type="text"
 					style="width: 180px;" value="${dateStart }"
 					onClick="new Calendar('1900',  '<%=Calendar.getInstance().get(1)%>', 0).show(this)"
-					class="ele_date" />
-				</td>
+					class="ele_date" /></td>
 				<td align="right">至</td>
 				<td align="left"><input name="txtDateEnd" type="text"
 					style="width: 180px;" value="${dateEnd }"
 					onClick=" new Calendar('1900',  '<%=Calendar.getInstance().get(1)%>', 0).show(this)"
-					class="ele_date" />
-				</td>
+					class="ele_date" /></td>
 				<td align="right">发送内容</td>
 				<td align="left"><input name="txtContext" type="text"
-					style="width: 180px;" value="${Context }" />
-				</td>
+					style="width: 180px;" value="${Context }" /></td>
 				<td><input id="btnQuery" type="submit" value="查询"
-					style="width: 100px;height: 30px;font-size: 16px" />
-				</td>
+					style="width: 100px;height: 30px;font-size: 16px" /></td>
 				<td><input id="btnReset" type="button" value="重置"
 					style="width: 100px;height: 30px;font-size: 16px"
-					onclick="formReset()" /></td>
+					onclick="formReset()" />
+				</td>
 				<td><input type="button" value="添加"
 					style="width: 100px;height: 30px;font-size: 16px"
-					onclick="window.location.href='add.action?param=0'" />
-				</td>
+					onclick="window.location.href='add.action?param=0'" /></td>
 			</tr>
 		</table>
 	</form>
@@ -172,13 +168,13 @@
 			<tr>
 				<td width="30">${msg.id}</td>
 				<td width="100"><fmt:formatDate type="date" value="${msg.date}"
-						dateStyle="default" /></td>
+						dateStyle="default" />
+				</td>
 				<td width="100">${msg.recName}</td>
 				<td width="100">${msg.telNum}</td>
 				<td width="100">${msg.context}</td>
 				<td><a href="edit.action?param=0&id=${msg.id}">编辑</a>&nbsp;&nbsp;<a
-					href="delete.action?id=${msg.id}">删除</a>
-				</td>
+					href="delete.action?id=${msg.id}" onclick="return confirm('确定要删除吗？')">删除</a></td>
 			</tr>
 		</c:forEach>
 	</table>
@@ -186,7 +182,21 @@
 		<tr>
 			<td colspan="6" align="center" bgcolor="#5BA8DE">共${pageBean.allRow
 				}条记录 共${pageBean.totalPage }页 当前第${pageBean.currentPage }页<br>
-			</td>
+				<c:choose>
+					<c:when test="${pageBean.currentPage ==1}"> 第一页 上一页 </c:when>
+					<c:otherwise>
+						<a href="query.action?page=1">第一页</a>
+						<a href="query.action?page=${pageBean.currentPage -1}">上一页</a>
+					</c:otherwise>
+				</c:choose> <c:choose>
+					<c:when test="${pageBean.currentPage ==pageBean.totalPage}">
+					下一页 最后一页
+					</c:when>
+					<c:otherwise>
+						<a href="query.action?page=${pageBean.currentPage +1}">下一页</a>
+						<a href="query.action?page=${pageBean.totalPage}">最后一页</a>
+					</c:otherwise>
+				</c:choose></td>
 		</tr>
 	</table>
 </body>
